requirement.txt
pyspark==3.5.1
venv-pack
pyyaml
ua-parser
user-agents
tldextract
pycountry
psycopg2-binary
